+++ packages/auth/src/providers/cognito/apis/signInWithCustomAuth.ts
@@ -21,7 +21,14 @@ import {
 	SignInWithCustomAuthInput,
 	SignInWithCustomAuthOutput,
 } from '../types';
+<<<<<<< HEAD
+import {
+	cleanActiveSignInState,
+	setActiveSignInState,
+} from '../../../client/utils/store';
+=======
 import { setActiveSignInState, signInStore } from '../utils/signInStore';
+>>>>>>> joonwonc/auth-resumable-signin
 import { cacheCognitoTokens } from '../tokenProvider/cacheTokens';
 import {
 	ChallengeName,