+++ packages/auth/src/providers/cognito/apis/signInWithCustomSRPAuth.ts
@@ -23,7 +23,14 @@ import {
 	SignInWithCustomSRPAuthInput,
 	SignInWithCustomSRPAuthOutput,
 } from '../types';
+<<<<<<< HEAD
+import {
+	cleanActiveSignInState,
+	setActiveSignInState,
+} from '../../../client/utils/store';
+=======
 import { setActiveSignInState, signInStore } from '../utils/signInStore';
+>>>>>>> joonwonc/auth-resumable-signin
 import { cacheCognitoTokens } from '../tokenProvider/cacheTokens';
 import {
 	ChallengeName,