+++ packages/auth/src/providers/cognito/apis/signInWithSRP.ts
@@ -27,10 +27,19 @@ import {
 	SignInWithSRPInput,
 	SignInWithSRPOutput,
 } from '../types';
+<<<<<<< HEAD
+import {
+	cleanActiveSignInState,
+	setActiveSignInState,
+} from '../../../client/utils/store';
+=======
 import { setActiveSignInState, signInStore } from '../utils/signInStore';
+>>>>>>> joonwonc/auth-resumable-signin
 import { cacheCognitoTokens } from '../tokenProvider/cacheTokens';
 import { tokenOrchestrator } from '../tokenProvider';
 import { dispatchSignedInHubEvent } from '../utils/dispatchSignedInHubEvent';
+
+import { resetAutoSignIn } from './autoSignIn';
 
 /**
  * Signs a user in
@@ -101,6 +110,8 @@ export async function signInWithSRP(
 
 			await dispatchSignedInHubEvent();
 
+			resetAutoSignIn();
+
 			return {
 				isSignedIn: true,
 				nextStep: { signInStep: 'DONE' },
@@ -112,7 +123,12 @@ export async function signInWithSRP(
 			challengeParameters: handledChallengeParameters as ChallengeParameters,
 		});
 	} catch (error) {
+<<<<<<< HEAD
+		cleanActiveSignInState();
+		resetAutoSignIn();
+=======
 		signInStore.dispatch({ type: 'RESET_STATE' });
+>>>>>>> joonwonc/auth-resumable-signin
 		assertServiceError(error);
 		const result = getSignInResultFromError(error.name);
 		if (result) return result;