+++ packages/core/src/clients/middleware/retry/index.ts
@@ -4,3 +4,4 @@
 export { RetryOptions, retryMiddlewareFactory } from './middleware';
 export { jitteredBackoff } from './jitteredBackoff';
 export { getRetryDecider } from './defaultRetryDecider';
+export { RetryDeciderOutput } from './types';