+++ packages/core/src/parseAmplifyOutputs.ts
@@ -5,10 +5,10 @@
 /* eslint-disable camelcase */
 
 /* Does not like exhaustive checks */
-/* eslint-disable no-case-declarations */
 
 import {
 	APIConfig,
+	APIEventsConfig,
 	APIGraphQLConfig,
 	GraphQLAuthMode,
 	ModelIntrospectionSchema,
@@ -22,6 +22,7 @@ import {
 	AmplifyOutputs,
 	AmplifyOutputsAnalyticsProperties,
 	AmplifyOutputsAuthProperties,
+	AmplifyOutputsCustomProperties,
 	AmplifyOutputsDataProperties,
 	AmplifyOutputsGeoProperties,
 	AmplifyOutputsNotificationsProperties,
@@ -87,12 +88,14 @@ function parseAuth(
 		oauth,
 		username_attributes,
 		standard_required_attributes,
+		groups,
 	} = amplifyOutputsAuthProperties;
 
 	const authConfig = {
 		Cognito: {
 			userPoolId: user_pool_id,
 			userPoolClientId: user_pool_client_id,
+			groups,
 		},
 	} as AuthConfig;
 
@@ -223,6 +226,28 @@ function parseData(
 	};
 }
 
+function parseCustom(
+	amplifyOutputsCustomProperties?: AmplifyOutputsCustomProperties,
+) {
+	if (!amplifyOutputsCustomProperties?.events) {
+		return undefined;
+	}
+
+	const { url, aws_region, api_key, default_authorization_type } =
+		amplifyOutputsCustomProperties.events;
+
+	const Events: APIEventsConfig = {
+		endpoint: url,
+		defaultAuthMode: getGraphQLAuthMode(default_authorization_type),
+		region: aws_region,
+		apiKey: api_key,
+	};
+
+	return {
+		Events,
+	};
+}
+
 function parseNotifications(
 	amplifyOutputsNotificationsProperties?: AmplifyOutputsNotificationsProperties,
 ): NotificationsConfig | undefined {
@@ -290,6 +315,14 @@ export function parseAmplifyOutputs(
 		resourcesConfig.API = parseData(amplifyOutputs.data);
 	}
 
+	if (amplifyOutputs.custom) {
+		const customConfig = parseCustom(amplifyOutputs.custom);
+
+		if (customConfig && 'Events' in customConfig) {
+			resourcesConfig.API = { ...resourcesConfig.API, ...customConfig };
+		}
+	}
+
 	if (amplifyOutputs.notifications) {
 		resourcesConfig.Notifications = parseNotifications(
 			amplifyOutputs.notifications,
@@ -342,18 +375,21 @@ function createBucketInfoMap(
 ): Record<string, BucketInfo> {
 	const mappedBuckets: Record<string, BucketInfo> = {};
 
-	buckets.forEach(({ name, bucket_name: bucketName, aws_region: region }) => {
-		if (name in mappedBuckets) {
-			throw new Error(
-				`Duplicate friendly name found: ${name}. Name must be unique.`,
-			);
-		}
+	buckets.forEach(
+		({ name, bucket_name: bucketName, aws_region: region, paths }) => {
+			if (name in mappedBuckets) {
+				throw new Error(
+					`Duplicate friendly name found: ${name}. Name must be unique.`,
+				);
+			}
 
-		mappedBuckets[name] = {
-			bucketName,
-			region,
-		};
-	});
+			mappedBuckets[name] = {
+				bucketName,
+				region,
+				paths,
+			};
+		},
+	);
 
 	return mappedBuckets;
 }