+++ packages/storage/src/providers/s3/types/options.ts
@@ -2,7 +2,11 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { StorageAccessLevel } from '@aws-amplify/core';
-import { SigningOptions } from '@aws-amplify/core/internals/aws-client-utils';
+import {
+	CredentialsProviderOptions,
+	SigningOptions,
+} from '@aws-amplify/core/internals/aws-client-utils';
+import { AWSCredentials } from '@aws-amplify/core/internals/utils';
 
 import { TransferProgressEvent } from '../../../types';
 import {
@@ -11,9 +15,32 @@ import {
 	StorageSubpathStrategy,
 } from '../../../types/options';
 
+/**
+ * @internal
+ */
+export type AWSTemporaryCredentials = Required<
+	Pick<
+		AWSCredentials,
+		'accessKeyId' | 'secretAccessKey' | 'sessionToken' | 'expiration'
+	>
+>;
+
+/**
+ * Async function returning AWS credentials for an API call. This function
+ * is invoked with S3 locations(bucket and path).
+ * If omitted, the global credentials configured in Amplify Auth
+ * would be used.
+ *
+ * @internal
+ */
+export type LocationCredentialsProvider = (
+	options?: CredentialsProviderOptions,
+) => Promise<{ credentials: AWSTemporaryCredentials }>;
+
 export interface BucketInfo {
 	bucketName: string;
 	region: string;
+	paths?: Record<string, Record<string, string[] | undefined>>;
 }
 
 export type StorageBucket = string | BucketInfo;
@@ -23,7 +50,13 @@ interface CommonOptions {
 	 * @default false
 	 */
 	useAccelerateEndpoint?: boolean;
+
 	bucket?: StorageBucket;
+
+	/**
+	 * The expected owner of the target bucket.
+	 */
+	expectedBucketOwner?: string;
 }
 
 /**
@@ -74,9 +107,9 @@ interface TransferOptions {
 /**
  * Input options type for S3 getProperties API.
  */
-/** @deprecated Use {@link GetPropertiesOptionsWithPath} instead. */
-export type GetPropertiesOptionsWithKey = ReadOptions & CommonOptions;
-export type GetPropertiesOptionsWithPath = CommonOptions;
+/** @deprecated Use {@link GetPropertiesWithPathOptions} instead. */
+export type GetPropertiesWithKeyOptions = ReadOptions & CommonOptions;
+export type GetPropertiesWithPathOptions = CommonOptions;
 
 /**
  * Input options type for S3 getProperties API.
@@ -84,25 +117,25 @@ export type GetPropertiesOptionsWithPath = CommonOptions;
 export type RemoveOptions = WriteOptions & CommonOptions;
 
 /**
- * @deprecated Use {@link ListAllOptionsWithPath} instead.
+ * @deprecated Use {@link ListAllWithPathOptions} instead.
  * Input options type with prefix for S3 list all API.
  */
-export type ListAllOptionsWithPrefix = StorageListAllOptions &
+export type ListAllWithPrefixOptions = StorageListAllOptions &
 	ReadOptions &
 	CommonOptions;
 
 /**
- * @deprecated Use {@link ListPaginateOptionsWithPath} instead.
+ * @deprecated Use {@link ListPaginateWithPathOptions} instead.
  * Input options type with prefix for S3 list API to paginate items.
  */
-export type ListPaginateOptionsWithPrefix = StorageListPaginateOptions &
+export type ListPaginateWithPrefixOptions = StorageListPaginateOptions &
 	ReadOptions &
 	CommonOptions;
 
 /**
  * Input options type with path for S3 list all API.
  */
-export type ListAllOptionsWithPath = Omit<
+export type ListAllWithPathOptions = Omit<
 	StorageListAllOptions,
 	'accessLevel'
 > &
@@ -113,7 +146,7 @@ export type ListAllOptionsWithPath = Omit<
 /**
  * Input options type with path for S3 list API to paginate items.
  */
-export type ListPaginateOptionsWithPath = Omit<
+export type ListPaginateWithPathOptions = Omit<
 	StorageListPaginateOptions,
 	'accessLevel'
 > &
@@ -150,9 +183,9 @@ export type GetUrlOptions = CommonOptions & {
 	contentType?: string;
 };
 
-/** @deprecated Use {@link GetUrlOptionsWithPath} instead. */
-export type GetUrlOptionsWithKey = ReadOptions & GetUrlOptions;
-export type GetUrlOptionsWithPath = GetUrlOptions;
+/** @deprecated Use {@link GetUrlWithPathOptions} instead. */
+export type GetUrlWithKeyOptions = ReadOptions & GetUrlOptions;
+export type GetUrlWithPathOptions = GetUrlOptions;
 
 /**
  * Input options type for S3 downloadData API.
@@ -161,9 +194,11 @@ export type DownloadDataOptions = CommonOptions &
 	TransferOptions &
 	BytesRangeOptions;
 
-/** @deprecated Use {@link DownloadDataOptionsWithPath} instead. */
-export type DownloadDataOptionsWithKey = ReadOptions & DownloadDataOptions;
-export type DownloadDataOptionsWithPath = DownloadDataOptions;
+/** @deprecated Use {@link DownloadDataWithPathOptions} instead. */
+export type DownloadDataWithKeyOptions = ReadOptions & DownloadDataOptions;
+export type DownloadDataWithPathOptions = DownloadDataOptions;
+
+export type UploadDataChecksumAlgorithm = 'crc-32';
 
 export type UploadDataOptions = CommonOptions &
 	TransferOptions & {
@@ -190,31 +225,51 @@ export type UploadDataOptions = CommonOptions &
 		 * @see https://docs.aws.amazon.com/AmazonS3/latest/userguide/UsingMetadata.html#UserMetadata
 		 */
 		metadata?: Record<string, string>;
+		/**
+		 * Enforces target key does not already exist in S3 before committing upload.
+		 * @default false
+		 */
+		preventOverwrite?: boolean;
+		/**
+		 * The algorithm used to compute a checksum for the object. Used to verify that the data received by S3
+		 * matches what was originally sent. Disabled by default.
+		 * @default undefined
+		 */
+		checksumAlgorithm?: UploadDataChecksumAlgorithm;
 	};
 
-/** @deprecated Use {@link UploadDataOptionsWithPath} instead. */
-export type UploadDataOptionsWithKey = WriteOptions & UploadDataOptions;
-export type UploadDataOptionsWithPath = UploadDataOptions;
+/** @deprecated Use {@link UploadDataWithPathOptions} instead. */
+export type UploadDataWithKeyOptions = WriteOptions & UploadDataOptions;
+export type UploadDataWithPathOptions = UploadDataOptions;
 
 /** @deprecated This may be removed in the next major version. */
-export type CopySourceOptionsWithKey = ReadOptions & {
+export type CopySourceWithKeyOptions = ReadOptions & {
 	/** @deprecated This may be removed in the next major version. */
 	key: string;
 	bucket?: StorageBucket;
+	notModifiedSince?: Date;
+	eTag?: string;
+	expectedBucketOwner?: string;
 };
 
 /** @deprecated This may be removed in the next major version. */
-export type CopyDestinationOptionsWithKey = WriteOptions & {
+export type CopyDestinationWithKeyOptions = WriteOptions & {
 	/** @deprecated This may be removed in the next major version. */
 	key: string;
 	bucket?: StorageBucket;
+	expectedBucketOwner?: string;
 };
 
 export interface CopyWithPathSourceOptions {
 	bucket?: StorageBucket;
+	notModifiedSince?: Date;
+	eTag?: string;
+	expectedBucketOwner?: string;
 }
+
 export interface CopyWithPathDestinationOptions {
 	bucket?: StorageBucket;
+	expectedBucketOwner?: string;
 }
 
 /**